+++ Realm/Realm/RealmConfiguration.cs
@@ -18,7 +18,9 @@
 
 using System;
 using System.IO;
+using System.Runtime.InteropServices;
 using Realms.Exceptions;
+using Realms.Native;
 using Realms.Schema;
 
 namespace Realms
@@ -49,6 +51,17 @@ namespace Realms
         public delegate void MigrationCallbackDelegate(Migration migration, ulong oldSchemaVersion);
 
         /// <summary>
+        /// A callback, invoked when opening a Realm for the first time during the life
+        /// of a process to determine if it should be compacted before being returned
+        /// to the user.
+        /// </summary>
+        /// <param name="totalBytes">Total file size (data + free space).</param>
+        /// <param name="bytesUsed">Total data size.</param>
+        /// <returns><c>true</c> to indicate that an attempt to compact the file should be made.</returns>
+        /// <remarks>The compaction will be skipped if another process is accessing it.</remarks>
+        public delegate bool ShouldCompactDelegate(ulong totalBytes, ulong bytesUsed);
+
+        /// <summary>
         /// Gets or sets a value indicating whether the database will be deleted if the <see cref="RealmSchema"/> 
         /// mismatches the one in the code. Use this when debugging and developing your app but never release it with
         /// this flag set to <c>true</c>.
@@ -71,6 +84,15 @@ namespace Realms
         /// to be migrated.
         /// </value>
         public MigrationCallbackDelegate MigrationCallback { get; set; }
+
+        /// <summary>
+        /// Gets or sets the compact on launch callback.
+        /// </summary>
+        /// <value>
+        /// The <see cref="ShouldCompactDelegate"/> that will be invoked when opening a Realm for the first time
+        /// to determine if it should be compacted before being returned to the user.
+        /// </value>
+        public ShouldCompactDelegate ShouldCompactOnLaunch { get; set; }
 
         /// <summary>
         /// Gets or sets the <see cref="RealmConfiguration"/> that is used when creating a new <see cref="Realm"/> without specifying a configuration.
@@ -137,6 +159,13 @@ namespace Realms
                 migration.PopulateConfiguration(ref configuration);
             }
 
+            if (ShouldCompactOnLaunch != null)
+            {
+                var handle = GCHandle.Alloc(ShouldCompactOnLaunch);
+                configuration.should_compact_callback = ShouldCompactOnLaunchCallback;
+                configuration.managed_should_compact_delegate = GCHandle.ToIntPtr(handle);
+            }
+
             var srPtr = IntPtr.Zero;
             try
             {
@@ -149,6 +178,21 @@ namespace Realms
 
             srHandle.SetHandle(srPtr);
             return new Realm(srHandle, this, schema);
+        }
+
+        [NativeCallback(typeof(ShouldCompactCallback))]
+        private static bool ShouldCompactOnLaunchCallback(IntPtr delegatePtr, ulong totalSize, ulong dataSize)
+        {
+            var handle = GCHandle.FromIntPtr(delegatePtr);
+            var compactDelegate = (ShouldCompactDelegate)handle.Target;
+            try
+            {
+                return compactDelegate(totalSize, dataSize);
+            }
+            finally
+            {
+                handle.Free();
+            }
         }
     }
 }