+++ wrappers/src/shared_realm_cs.cpp
@@ -100,6 +100,17 @@ REALM_EXPORT SharedRealm* shared_realm_open(Configuration configuration, SchemaO
             };
         }
         
+        if (configuration.managed_should_compact_delegate) {
+#ifndef _WIN32
+            config.should_compact_on_launch_function = [&configuration](uint64_t total_bytes, uint64_t used_bytes) {
+                return configuration.should_compact_callback(configuration.managed_should_compact_delegate, total_bytes, used_bytes);
+            };
+#else
+            throw std::logic_error("Compact isn't supported on Windows yet.");
+#endif
+            
+        }
+        
         return new SharedRealm{Realm::get_shared_realm(config)};
     });
 }
@@ -202,8 +213,12 @@ REALM_EXPORT size_t shared_realm_refresh(SharedRealm* realm, NativeException::Ma
 
 REALM_EXPORT bool shared_realm_compact(SharedRealm* realm, NativeException::Marshallable& ex)
 {
-    return handle_errors(ex, [&]() {
+    return handle_errors(ex, [&]() -> bool {
+#ifndef _WIN32
         return (*realm)->compact();
+#else
+        throw std::logic_error("Compact isn't supported on Windows yet.");
+#endif
     });
 }