+++ wrappers/src/shared_realm_cs.hpp
@@ -46,6 +46,9 @@ struct Configuration
     
     bool (*migration_callback)(realm::SharedRealm* old_realm, realm::SharedRealm* new_realm, SchemaForMarshaling, uint64_t schema_version, void* managed_migration_handle);
     void* managed_migration_handle;
+    
+    bool (*should_compact_callback)(void* managed_config_handle, uint64_t total_size, uint64_t data_size);
+    void* managed_should_compact_delegate;
 };
 
 namespace realm {